implemet some changes
