+++ apps/frontend/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import MesaChainSidebar from "../components/SideBar";
+import MesaChainSidebar from "../components/MesaChainSidebar";
 import "./globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
