+++ apps/frontend/app/page.tsx
@@ -1,7 +1,7 @@
 export default function Home() {
   console.log("home page")
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24 bg-red-500">
+    <main className="flex min-h-screen flex-col items-center justify-between p-24">
       <h1>Welcome</h1>
     </main>
   );
